[CMA-157] fix : 나의 프로필 id를 0으로 가져오는 에러 수정

+++ app/src/main/java/com/climus/climeet/data/remote/MainApi.kt
@@ -498,7 +498,7 @@ interface MainApi {
     ): Response<GetMyShortsCommentsResponse>
 
     @GET("/api/manager/gym-id")
-    suspend fun getAdminGymId(): Response<Long>
+    suspend fun getAdminGymId(): Response<Int>
 
     @PATCH("/api/gyms/service")
     suspend fun updateGymService(
